+++ src/Eventuous.Subscriptions.SqlStreamStore.MsSql/Eventuous.Subscriptions.SqlStreamStore.MsSql.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="SqlStreamStore.Mysql" Version="1.2.0" />
+    <PackageReference Include="SqlStreamStore.MsSql" Version="1.2.0" />
   </ItemGroup>
 
   <ItemGroup>
