+++ src/Eventuous.Subscriptions.SqlStreamStore.MsSql/StreamSusbscription.cs
@@ -6,13 +6,13 @@ using SqlStreamStore;
 using SqlStreamStore.Streams;
 using Eventuous.Subscriptions.SqlStreamStore;
 
-namespace Eventuous.Subscriptions.SqlStreamStore.MySql
+namespace Eventuous.Subscriptions.SqlStreamStore.MsSql
 {
     /// <summary>
-    /// Producer for SqlStreamStore (https://sqlstreamstore.readthedocs.io) which has a MySQL database as the event data store.
+    /// Producer for SqlStreamStore (https://sqlstreamstore.readthedocs.io) which has a MsSQL database as the event data store.
     /// </summary>
     [PublicAPI]
-    public class AllStreamSubscription : SqlStreamStore.AllStreamSubscription
+    public class StreamSubscription : SqlStreamStore.StreamSubscription
     {
         /// <summary>
         /// Creates SqlStreamStore catch-up subscription service for $all
@@ -25,17 +25,23 @@ namespace Eventuous.Subscriptions.SqlStreamStore.MySql
         /// <param name="loggerFactory">Optional: logger factory</param>
         /// <param name="measure">Optional: gap measurement for metrics</param>
 
-        public AllStreamSubscription(
-            MySqlStreamStore            mySqlStore,
+        public StreamSubscription(
+            MsSqlStreamStoreV3          msSqlStore,
+            string                      streamName,
             string                      subscriptionId,
             ICheckpointStore            checkpointStore,
             IEnumerable<IEventHandler>  eventHandlers,
             IEventSerializer?           eventSerializer = null,
             ILoggerFactory?             loggerFactory   = null,
-            ISubscriptionGapMeasure?    measure         = null
+            ISubscriptionGapMeasure?    measure         = null,
+            bool                        throwOnError    = false
         ) : base(
-            Ensure.NotNull(mySqlStore, nameof(mySqlStore)),
-            new AllStreamSubscriptionOptions { SubscriptionId = subscriptionId},
+            Ensure.NotNull(msSqlStore, nameof(msSqlStore)),
+            new StreamSubscriptionOptions { 
+                StreamName = streamName,
+                SubscriptionId = subscriptionId,
+                ThrowOnError = throwOnError
+            },
             checkpointStore,
             eventHandlers,
             eventSerializer,
